% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality_neighborhood.R
\name{centrality_neighborhood}
\alias{centrality_neighborhood}
\title{Neighborhood centrality}
\usage{
centrality_neighborhood(graph, actors, layers)
}
\arguments{
\item{graph}{A graph of class tbl_graph}

\item{actors}{An actor or a vector of actors. Defaults to all actors in the graph.}

\item{layers}{A layer or a vector of layers. Defaults to all layers of the graph}
}
\description{
Calculate neighborhood centrality for a set of actors over a set of layers
}
