% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidesc.R
\name{compare_layers}
\alias{compare_layers}
\title{A function for generating a list of descriptive measurements for the structure of a multilayer network}
\usage{
compare_layers(graph, var)
}
\description{
A function for generating a list of descriptive measurements for the structure of a multilayer network
}
