% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connective_redundancy.R
\name{connective_redundancy}
\alias{connective_redundancy}
\title{Connective Redundancy}
\usage{
connective_redundancy(graph, actors, layers)
}
\arguments{
\item{graph}{A graph object of the tbl_graph type}

\item{actors}{An actor or a vector of actors, stored in the "id" column.}

\item{layers}{A layer or a vector of layers, stored in the "name" column.}
}
\description{
Calculate the connective redundancy of a set of actors over a set of layers within a multilayer graph.
}
