% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_correlation.R
\name{dimension_correlation}
\alias{dimension_correlation}
\title{dimension correlation}
\usage{
dimension_correlation(graph, weighted = FALSE)
}
\arguments{
\item{graph}{A graph of class tbl_graph}

\item{weighted}{Logical value if weight of the edges of the dimensions is considered or not. Defaults to unweighted network.}
}
\description{
Calculate the correlation of the dimensions of a graph. The output is a correlation matrix of the layers.
}
