% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_matlist.R
\name{dimension_matlist}
\alias{dimension_matlist}
\title{dimension correlation}
\usage{
dimension_matlist(graph, actors, dimensions, weighted = FALSE)
}
\arguments{
\item{graph}{A graph of class tbl_graph}

\item{actors}{An actor or a vector of actors. Defaults to all actors in the graph.}

\item{dimensions}{The Dimensions of the edges that should be taken into account. A variable of the edgelist.}

\item{weighted}{Logical value if weight of the edges of the dimensions is considered or not. Defaults to unweighted network.}
}
\description{
Creates a list containing the adjacency matrices for each layer of the network. This can be used for further analyses of the layers (for example their correlation).
}
