% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edges_decor}
\alias{edges_decor}
\title{Edges between 22 archaeological sites in West Africa.}
\format{
A data frame with 661 rows and 6 variables:
\describe{
  \item{from}{originating node}
  \item{to}{receiving node}
  \item{weight}{weight of the edge}
  \item{type}{archaeological find class}
  \item{subtype}{type of observation}
  \item{name}{observation, this is the network layer}
}
}
\usage{
edges_decor
}
\description{
An edgelist dataset containing the connections between 22 archaeological sites in West Africa
}
\keyword{datasets}
