% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten}
\alias{flatten}
\title{A function to flatten a multilayer network into a single-layer network, see Dickison et al. 2016, p.39}
\usage{
flatten(graph, weight)
}
\description{
A function to flatten a multilayer network into a single-layer network, see Dickison et al. 2016, p.39
}
