% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidesc.R
\name{multidesc}
\alias{multidesc}
\title{A function for generating a list of descriptive measurements for the structure of a multilayer network}
\usage{
multidesc(graph)
}
\description{
A function for generating a list of descriptive measurements for the structure of a multilayer network
}
