% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nodes_decor}
\alias{nodes_decor}
\title{Nodes of a network of 22 archaeological sites in West Africa}
\format{
A data frame with 22 rows and 4 columns
\describe{
  \item{id}{node id}
  \item{label}{name of site}
  \item{lat}{geographical latitude}
  \item{lon}{geographical longitude}
}
}
\usage{
nodes_decor
}
\description{
A nodelist dataset containing 22 archaeological sites in West Africa as the nodes of a graph.
}
\keyword{datasets}
