% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_correlation.R
\name{dimension_correlation}
\alias{dimension_correlation}
\title{dimension correlation}
\usage{
dimension_correlation(matlist, pvalue = FALSE)
}
\arguments{
\item{pvalue}{Locial value if pvalue should be computed. If it is then the output is a list with 2 matrices.}

\item{graph}{A graph of class tbl_graph}
}
\description{
Calculate the correlation of the dimensions of a graph. The output is a correlation matrix of the layers.
}
