% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_correlation_node.R
\name{dimension_correlation_node}
\alias{dimension_correlation_node}
\title{dimension correlation}
\usage{
dimension_correlation_node(
  graph,
  actors,
  dimensions,
  weighted = FALSE,
  pvalue = FALSE
)
}
\arguments{
\item{graph}{A graph of class tbl_graph}

\item{weighted}{Logical value if weight of the edges of the dimensions is considered or not. Defaults to unweighted network.}
}
\description{
Calculate the correlation of the dimensions of a graph based on the neighborhood centrality of the nodes in each layer. The output is a correlation matrix of the layers.
}
