% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten}
\alias{flatten}
\title{Flatten a multilayer network}
\usage{
flatten(graph, weighted = FALSE)
}
\arguments{
\item{graph}{A graph of the tbl-graph class}

\item{weighted}{should the flattening take edge weights into account (TRUE/FALSE)? Defaults to unweighted edges. Note that weights must be stored in an edge variable named "weight".}
}
\description{
A function to flatten a multilayer network into a single-layer network, see Dickison et al. 2016, p.39
}
